make frequency float +  calc biquad in set_eq_gain

+++ components/tas58xx/tas58xx_helpers.h
@@ -21,6 +21,6 @@ int32_t gain_to_f9_23_(int8_t gain);
 
 BiquadCoefficients butterworth2_(uint16_t fs, uint16_t fc, Butterworth2Type type);
 
-BiquadCoefficients equalizer_qfactor_calc(uint32_t sample_rate, uint16_t frequency, int16_t gain, float qFactor);
+BiquadCoefficients equalizer_qfactor_calc(uint32_t sample_rate, float frequency, int16_t gain, float qFactor);
 
 }  // namespace esphome::tas58xx_helpers
